+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContext.cs
@@ -36,6 +36,8 @@ namespace UniGLTF
         MaterialFactory m_materialFactory;
         public MaterialFactory MaterialFactory => m_materialFactory;
 
+        public readonly GltfMaterialImporter GltfMaterialImporter = new GltfMaterialImporter();
+
         TextureFactory m_textureFactory;
         public TextureFactory TextureFactory => m_textureFactory;
 
@@ -46,7 +48,7 @@ namespace UniGLTF
         {
             m_parser = parser;
             m_textureFactory = new TextureFactory(externalObjectMap);
-            m_materialFactory = new MaterialFactory(m_parser, externalObjectMap);
+            m_materialFactory = new MaterialFactory(externalObjectMap);
         }
 
         #region Source
@@ -111,7 +113,7 @@ namespace UniGLTF
 
             using (MeasureTime("LoadMaterials"))
             {
-                await m_materialFactory.LoadMaterialsAsync(m_awaitCaller, m_textureFactory.GetTextureAsync);
+                await LoadMaterialsAsync();
             }
 
             var meshImporter = new MeshImporter();
@@ -169,6 +171,24 @@ namespace UniGLTF
             }
 
             await OnLoadModel(m_awaitCaller, MeasureTime);
+        }
+
+        public async Task LoadMaterialsAsync()
+        {
+            if (m_parser.GLTF.materials == null || m_parser.GLTF.materials.Count == 0)
+            {
+                // no material. work around.
+                var param = GltfMaterialImporter.CreateParam(m_parser, 0);
+                var material = await MaterialFactory.LoadAsync(param, TextureFactory.GetTextureAsync);
+            }
+            else
+            {
+                for (int i = 0; i < m_parser.GLTF.materials.Count; ++i)
+                {
+                    var param = GltfMaterialImporter.CreateParam(m_parser, i);
+                    var material = await MaterialFactory.LoadAsync(param, TextureFactory.GetTextureAsync);
+                }
+            }
         }
 
         protected virtual Task OnLoadModel(IAwaitCaller awaitCaller, Func<string, IDisposable> MeasureTime)
