+++ Assets/UniGLTF/Runtime/UniGLTF/IO/MaterialIO/GltfMaterialImporter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14bc3a4ce3e24924a9ae8bb1654a1b7b
+guid: 06d7ddbd9b8b38544a74013a6992210b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
