+++ Assets/UniGLTF/Runtime/UniGLTF/IO/MeshImporter.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using UnityEngine;
-
+using VRMShaders;
 
 namespace UniGLTF
 {
