+++ Assets/VRM/Runtime/IO/MToonMaterialImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35c90d5d3fa706b4f87a92ce4dc59008
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
