+++ Assets/VRM/Tests/MToonTest.cs
@@ -29,7 +29,7 @@ namespace VRM
             var vrmMaterial = VRMMaterialExporter.CreateFromMaterial(srcMaterial, textureManager);
             Assert.AreEqual(vrmMaterial.vectorProperties["_MainTex"], new float[]{0.3f, 0.2f, 0.5f, 0.6f});
             
-            var materialImporter = new VRMMaterialImporter(new System.Collections.Generic.List<glTF_VRM_Material>{ vrmMaterial });
+            var materialImporter = new MToonMaterialImporter(new System.Collections.Generic.List<glTF_VRM_Material>{ vrmMaterial });
         }
     }
 }
