+++ Assets/VRMShaders/Runtime/MaterialImportParam.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 77db57076c0650c469c3ff0d4853ce21
-timeCreated: 1533624711
-licenseType: Free
+guid: 0844821341efaf644b91bff72f813a2c
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
